+++ lib/auth.js
@@ -11,8 +11,5 @@ export const authOptions = {
             clientId: process.env.GOOGLE_ID,
             clientSecret: process.env.GOOGLE_SECRET,
         }),
-    ],
-    pages: {
-        signIn: '/signin',
-    },
+    ]
 };
